+++ src/surface.ts
@@ -1,6 +1,8 @@
 import fs from 'fs';
 import { GeometryControls } from 'molstar/lib/commonjs/extensions/geo-export/controls';
 import { VisualQuality, VisualQualityNames } from 'molstar/lib/commonjs/mol-geo/geometry/base';
+import { type GraphicsRenderObject } from 'molstar/lib/commonjs/mol-gl/render-object';
+import { Unit } from 'molstar/lib/commonjs/mol-model/structure';
 import { Download, ParseCif } from 'molstar/lib/commonjs/mol-plugin-state/transforms/data';
 import { ModelFromTrajectory, StructureComponent, StructureFromModel, TrajectoryFromMmCif } from 'molstar/lib/commonjs/mol-plugin-state/transforms/model';
 import { StructureRepresentation3D } from 'molstar/lib/commonjs/mol-plugin-state/transforms/representation';
@@ -73,6 +75,10 @@ export async function computeSurface(plugin: PluginContext, structureRef: Struct
             },
         })
         .commit();
+    return {
+        structure: surface.data!.sourceData,
+        meshes: surface.data!.repr.renderObjects.filter(obj => obj.type === 'mesh') as GraphicsRenderObject<'mesh'>[],
+    };
 }
 
 export async function exportGeometry(plugin: PluginContext, filename: string) {
@@ -81,4 +87,80 @@ export async function exportGeometry(plugin: PluginContext, filename: string) {
     const data = await geo.exportGeometry();
     const buffer = await data.blob.arrayBuffer();
     fs.writeFileSync(filename, Buffer.from(buffer));
+}
+
+export function getSurfaceMetadata(surface: Awaited<ReturnType<typeof computeSurface>>) {
+    const unitOffsets = getUnitOffsets(surface.structure.units);
+    const group_properties = getGroupProps(surface.structure.units);
+    const vertex_properties = getVertexProps(surface.meshes, unitOffsets);
+    return { group_properties, vertex_properties };
+}
+
+/** Return group offsets for struct units, i.e. j-th atom in i-th unit has total index `getUnitOffsets(units)[i] + j`. */
+function getUnitOffsets(units: readonly Unit[]) {
+    const offsets: Record<number, number> = {};
+    let offset = 0;
+    for (let iUnit = 0; iUnit < units.length; iUnit++) {
+        offsets[iUnit] = offset;
+        offset += units[iUnit].elements.length;
+    }
+    return offsets;
+}
+
+/** Return group properties (value per group) */
+function getGroupProps(units: readonly Unit[]) {
+    const props = {
+        atom_id: [] as number[],
+        label_atom_id: [] as string[],
+        label_comp_id: [] as string[],
+        label_seq_id: [] as number[],
+        label_asym_id: [] as string[],
+        auth_asym_id: [] as string[],
+        label_entity_id: [] as string[],
+    };
+    for (const unit of units) {
+        const h = unit.model.atomicHierarchy;
+        const conf = unit.model.atomicConformation;
+        for (let i = 0; i < unit.elements.length; i++) { // eslint-disable-line @typescript-eslint/prefer-for-of
+            const iElement = unit.elements[i];
+            const iResidue = h.residueAtomSegments.index[iElement];
+            const iChain = h.chainAtomSegments.index[iElement];
+
+            props.atom_id.push(conf.atomId.value(iElement));
+            props.label_atom_id.push(h.atoms.label_atom_id.value(iElement));
+            props.label_comp_id.push(h.atoms.label_comp_id.value(iElement));
+            props.label_seq_id.push(h.residues.label_seq_id.value(iResidue));
+            props.label_asym_id.push(h.chains.label_asym_id.value(iChain));
+            props.auth_asym_id.push(h.chains.auth_asym_id.value(iChain));
+            props.label_entity_id.push(h.chains.label_entity_id.value(iChain));
+        }
+    }
+    return props;
+}
+
+
+/** Return vertex properties (value per vertex) */
+function getVertexProps(meshes: GraphicsRenderObject<'mesh'>[], unitOffsets: Record<number, number>) {
+    const props = {
+        group_index: [] as number[],
+    };
+    for (let iMesh = 0; iMesh < meshes.length; iMesh++) {
+        const mesh = meshes[iMesh];
+        const groups = mesh.values.aGroup.ref.value;
+        for (let iVertex = 0; iVertex < groups.length; iVertex++) { // eslint-disable-line @typescript-eslint/prefer-for-of
+            const groupIndexWithinMesh = groups[iVertex];
+            const groupIndex = unitOffsets[iMesh] + groupIndexWithinMesh;
+            props.group_index.push(groupIndex);
+        }
+    }
+    return props;
+}
+
+/** JSON.stringify object of objects of arrays, with putting each array on separate line (dumb implementation but enough for what we need). */
+export function niceJsonStringify(data: Record<string, Record<string, any[]>>) {
+    return JSON.stringify(data)
+        .replace(/("\w*":)\{/g, '\n  $1 {')
+        .replace(/("\w*":)\[/g, '\n    $1 [')
+        .replace(/\]\}/g, ']\n  }')
+        .replace(/\}\}/g, '}\n}\n');
 }